Sync SchemaTrait [skip ci]
